pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "escape-sdk"
version = "1.0.2"
description = "OSRS Bot Development SDK with game-native structure"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Escape Team"}
]
keywords = ["osrs", "runescape", "bot", "automation", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "msgpack>=1.0.0",
    "inotify-simple>=1.3.5",
    "pyautogui>=0.9.54",
    "python-xlib>=0.33",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "loguru>=0.7.0",
    "ruff>=0.14.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.8.4",
    "pre-commit>=3.5.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/escape/escape"
Repository = "https://github.com/escape/escape"
Issues = "https://github.com/escape/escape/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["escape*"]
exclude = [
    "tests*",
    "docs*",
]

[tool.setuptools.package-data]
escape = [
    "py.typed",
    "_internal/**/*.pyi",
]

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.ruff.lint]
# Enable pycodestyle (E, W), Pyflakes (F), and isort (I)
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]

# We'll ignore N802 (function name should be lowercase) since we want camelCase
ignore = [
    "N802",  # Function name should be lowercase (we use camelCase)
    "N806",  # Variable in function should be lowercase (allow camelCase)
    "N803",  # Argument name should be lowercase (allow camelCase)
    "N815",  # Variable in class scope should not be mixedCase (allow camelCase in dataclasses)
    "N816",  # Variable in global scope should not be mixedCase (allow camelCase for module functions)
    "F401",  # Imported but unused (scripts use imports for testing)
    "F821",  # Undefined name (many are dynamically generated at runtime)
    "F822",  # Undefined name in __all__ (dynamically generated)
    "UP006", # Use `dict` instead of `Dict` (for consistency with existing code)
    "UP035", # Deprecated typing imports (TODO: migrate to modern type hints)
    "B905",  # zip() without strict parameter (TODO: add strict=True where appropriate)
    "TC004", # Move import out of TYPE_CHECKING (needed at runtime in some cases)
    "E501",  # Line too long (already handled by formatter)
    "SIM102", # Use single if statement (nested ifs are often more readable)
    "SIM108", # Use ternary operator (if-else blocks are often more readable)
    "SIM115", # Use context manager (not always applicable for mmap file handles)
    "SIM117", # Use single with statement (nested with is often clearer)
    "B007",  # Loop control variable not used (sometimes intentional)
    "B023",  # Function definition does not bind loop variable (false positives)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["escape"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.4.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=escape",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["escape"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
